% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfilePlot.R
\name{ProfilePlot}
\alias{ProfilePlot}
\title{Plot the phylogenetic profiles}
\usage{
ProfilePlot(profile, dend, ...)
}
\arguments{
\item{profile}{a data frame or matrix contains the phylogenetic profiles.
Each column is a profile. The row names must match the tip labels of the tree.}

\item{dend}{the dendrogram that will be plotted on the left side. See `TreeToDen()`.}

\item{...}{some other parameters available for `heatmap.2()`, such as `main`, `xlab`, `labRow`...}
}
\description{
Plot the phylogenetic profiles along with the phylogenetic tree.
This function is based on `heatmap.2()` from `gplots` package.
}
\examples{
d <- TreeToDend(rtree(50))
p <-  matrix(sample(0:1,100, replace=T), nrow=50)
rownames(p) <- labels(d)
ProfilePlot(p, d, main="Plot of two profiles")
}
