% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeToDen.R
\name{TreeToDend}
\alias{TreeToDend}
\title{Convert tree to dendrogram}
\usage{
TreeToDend(phytree)
}
\arguments{
\item{tree}{an object of class "phylo"}
}
\value{
a dendrogram as an input for `ProfilePlot()` function
}
\description{
Convert the non-ultrametric tree to a dendrogram for `ProfilePlot()` to plot the profiles using
the `chronos()` function from `ape` package.
}
\examples{
#library(ape)
# create a random 50-tip tree
t = rtree(50)
plot(t)
d = TreeToDend(t)
plot(d)
}
